requirements.txt
httpx
yarl
tkinter
customtkinter
Pillow
CTkMessagebox